Fix Codeclimate issues with GridSenderTest
