+++ app/src/androidTest/java/ch/epfl/sdp/contamination/GridSenderTest.java
@@ -22,8 +22,6 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 import ch.epfl.sdp.Callback;
 import ch.epfl.sdp.R;
@@ -255,139 +253,5 @@ public class GridSenderTest {
         aMap.put(c1, 2);
 
         assertThat(aMap.containsKey(c2), is(true));
-    }
-
-    private Map<Carrier, Boolean> getBackSliceData(Location somewhere, Date rightNow) throws Throwable {
-        Map<Carrier, Boolean> result = new ConcurrentHashMap<>();
-
-        AtomicBoolean done = new AtomicBoolean();
-        done.set(false);
-
-        mActivityRule.runOnUiThread(() -> mActivityRule.getActivity().getService().getReceiver().getUserNearby(somewhere, rightNow, people -> {
-            for (Carrier c : people) {
-                result.put(c, false);
-            }
-            done.set(true);
-        }));
-
-        while (!done.get()) { } // Busy wait
-
-        return result;
-    }
-
-    @Test
-    public void dataReallyComeAndGoFromServer() throws Throwable {
-        // The following test uses the actual Firestore
-        resetRealSenderAndReceiver();
-
-        Carrier aFakeCarrier = new Layman(Carrier.InfectionStatus.UNKNOWN, 0.2734f);
-        Date rightNow = new Date(System.currentTimeMillis());
-
-        mActivityRule.getActivity().getService().getSender().registerLocation(
-                aFakeCarrier,
-                buildLocation(12, 73),
-                rightNow,
-                exchangeSucceeded,
-                exchangeFailed);
-
-        Thread.sleep(2000);
-
-        Map<Carrier, Boolean> result = getBackSliceData(buildLocation(12, 73), rightNow);
-
-        assertThat(result.size(), is(1));
-        assertThat(result.containsKey(aFakeCarrier), is(true));
-    }
-
-    private Map<Carrier, Integer> getBackRangeData(Location somewhere, Date rangeStart, Date rangeEnd) throws Throwable {
-        AtomicBoolean done = new AtomicBoolean(false);
-
-        Map<Carrier, Integer> result = new ConcurrentHashMap<>();
-
-        // Get data back
-        mActivityRule.runOnUiThread(() -> mActivityRule.getActivity().getService().getReceiver().getUserNearbyDuring(somewhere, rangeStart, rangeEnd, contactFrequency -> {
-            result.putAll(contactFrequency);
-            done.set(true);
-        }));
-
-        while (!done.get()) { } // Busy wait
-
-        return result;
-    }
-
-    @Test
-    public void complexQueriesComeAndGoFromServer() throws Throwable {
-        // The following test uses the actual Firestore
-
-        resetRealSenderAndReceiver();
-
-        Carrier aFakeCarrier = new Layman(Carrier.InfectionStatus.UNKNOWN, 0.2734f);
-        Carrier trulyHealthy = new Layman(Carrier.InfectionStatus.IMMUNE, 0f);
-
-        Location somewhereInTheWorld = buildLocation(12, 73);
-
-        Date rightNow = new Date(System.currentTimeMillis());
-        Date aLittleLater = new Date(rightNow.getTime() + 10);
-
-        mActivityRule.getActivity().getService().getSender().registerLocation(
-                aFakeCarrier,
-                somewhereInTheWorld,
-                rightNow,
-                exchangeSucceeded,
-                exchangeFailed);
-        mActivityRule.getActivity().getService().getSender().registerLocation(
-                trulyHealthy,
-                somewhereInTheWorld,
-                aLittleLater);
-
-        Thread.sleep(1000);
-
-        onView(withId(R.id.exchange_status)).check(matches(withText("EXCHANGE Succeeded")));
-
-        Map<Carrier, Integer> result = getBackRangeData(somewhereInTheWorld, rightNow, aLittleLater);
-
-        assertThat(result.size(), is(2));
-        assertThat(result.containsKey(aFakeCarrier), is(true));
-        assertThat(result.containsKey(trulyHealthy), is(true));
-        assertThat(result.get(aFakeCarrier), is(1));
-        assertThat(result.get(trulyHealthy), is(1));
-
-    }
-
-    @Test
-    public void repetitionsOfSameCarrierAreDetected() throws Throwable {
-        // The following test uses the actual Firestore
-
-        resetRealSenderAndReceiver();
-
-        Carrier aFakeCarrier = new Layman(Carrier.InfectionStatus.UNKNOWN, 0.2734f);
-
-        Location somewhereInTheWorld = buildLocation(12, 73);
-
-        Date rightNow = new Date(System.currentTimeMillis());
-        Date aLittleLater = new Date(rightNow.getTime() + 1000);
-
-        mActivityRule.runOnUiThread(() -> {
-            mActivityRule.getActivity().getService().getSender().registerLocation(
-                    aFakeCarrier,
-                    somewhereInTheWorld,
-                    rightNow,
-                    exchangeSucceeded,
-                    exchangeFailed);
-            mActivityRule.getActivity().getService().getSender().registerLocation(
-                    aFakeCarrier,
-                    somewhereInTheWorld,
-                    aLittleLater);
-        });
-
-        Thread.sleep(1000);
-
-        onView(withId(R.id.exchange_status)).check(matches(withText("EXCHANGE Succeeded")));
-
-        Map<Carrier, Integer> result = getBackRangeData(somewhereInTheWorld, rightNow, aLittleLater);
-
-        assertThat(result.size(), is(1));
-        assertThat(result.containsKey(aFakeCarrier), is(true));
-        assertThat(result.get(aFakeCarrier), is(2));
-
     }
 }
